CHORES: refactor global_var_error to property_error
